refactor: Adds folders for more clarity
