+++ src/index.ts
@@ -1,14 +1,14 @@
-import { generateSummaryTableHtml } from './generateSummaryTableHtml.js';
-import { parseVitestJsonFinal, parseVitestJsonSummary } from './parseJsonReports.js';
+import { FileCoverageMode } from './inputs/FileCoverageMode.js'
+import { generateFileCoverageHtml } from './report/generateFileCoverageHtml.js';
+import { generateHeadline } from './report/generateHeadline.js';
+import { generateSummaryTableHtml } from './report/generateSummaryTableHtml.js';
+import { getPullChanges } from './inputs/getPullChanges.js';
+import { parseVitestJsonFinal, parseVitestJsonSummary } from './inputs/parseJsonReports.js';
+import { readOptions } from './inputs/options.js';
+import { RequestError } from '@octokit/request-error'
 import { writeSummaryToPR } from './writeSummaryToPR.js';
 import * as core from '@actions/core';
 import * as github from '@actions/github';
-import { RequestError } from '@octokit/request-error'
-import { generateFileCoverageHtml } from './generateFileCoverageHtml.js';
-import { getPullChanges } from './getPullChanges.js';
-import { FileCoverageMode } from './FileCoverageMode.js'
-import { readOptions } from './options.js';
-import { generateHeadline } from './generateHeadline.js';
 
 const run = async () => {
 	const {
