+++ src/inputs/getViteConfigPath.test.ts
