+++ src/report/generateFileCoverageHtml.test.ts
@@ -1,11 +1,11 @@
-import { createJsonFinalEntry } from './types/JsonFinalMockFactory';
+import { createJsonFinalEntry } from '../types/JsonFinalMockFactory';
 import { generateFileCoverageHtml } from './generateFileCoverageHtml';
-import { getTableLine } from '../test/queryHelper';
-import { JsonFinal } from './types/JsonFinal';
-import { JsonSummary } from './types/JsonSummary';
-import { createMockCoverageReport, createMockJsonSummary, createMockReportNumbers } from './types/JsonSummaryMockFactory';
+import { getTableLine } from '../../test/queryHelper';
+import { JsonFinal } from '../types/JsonFinal';
+import { JsonSummary } from '../types/JsonSummary';
+import { createMockCoverageReport, createMockJsonSummary, createMockReportNumbers } from '../types/JsonSummaryMockFactory';
 import { describe, it, expect } from 'vitest';
-import { FileCoverageMode } from './FileCoverageMode';
+import { FileCoverageMode } from '../inputs/FileCoverageMode';
 import * as path from 'path';
 
 const workspacePath = process.cwd();
