+++ src/report/generateHeadline.ts
